add README, migrations, more tests, better logic
This commit adds a README to this project to provide a high-level
description of what it does.

It also adds some additional unit tests to verify that block detection
works with a variety of nested block types. The logic for block
detection has also been cleaned up a bit to make it easier to test.

Missing migrations are also added here.

The set of Python/Django/Wagtail versions to test against has been
updated to test against Wagtail 1.12 and Python 3.6.