fix broken pagination in search results
The changes in #11 to add Django/Wagtail 2.x support made a slight
change to how the wagtail-inventory search URL is defined:

https://github.com/cfpb/wagtail-inventory/commit/5900a7993ba5c086206d644f32d144192678effc#diff-e5cdc1d4255de591a575368291bee269L30

Unfortunately, this broke pagination of wagtail-inventory results in the
admin. In version 0.4.2 and previous, the admin page would erroneously
load with any URL, e.g. /admin/inventory/anything?foo=bar... Along with
this, the use of replace_page_in_query to generate page URLs wasn't
creating proper URLs. It so happened that the combination of these two
things made working (if ugly) URLs for pagination, but the change in 0.5
broken this.

This commit fixes pagination by leveraging the Wagtail pagination_nav
template, which handles generating pagination URLs for us.