Add missing chevron icons to select dropdowns
As reported in issue 39, the select dropdowns in the admin search form
should have chevron icons to indicate that they can be selected.

This commit adds those icons by leveraging the built-in field template
in the Wagtail admin (wagtailadmin/shared/field.html). This also allows
us to simplify the form and HTML code a bit.