Add support for Wagtail 3.0 and Django 4.0
This commit adds support for Wagtail version 3.0 and Django version
4.0, while still maintaining support for all previous versions.

To comply the Django 3.2 recommendation around use of BigAutoField for
model primary keys, a migration has been added that alters the type
of the PageBlock primary key to this type.

Automatic testing of Wagtail 2.16 has been dropped in favor of testing
against Wagtail 3.0.