Convert inventory view to a Wagtail report
This change converts the inventory view to a report. As a result, the "Block Inventory" that used to appear under "Settings" in the Wagtail admin will move to "Reports", and the UI for filtering will change. This provides us with built-in support for exporting the inventory list to an Excel file or a CSV.

It also uses Wagtail's filtering for reports, via django-filters. The UI now provides two selections, one of blocks to include in the report, and one of blocks to exclude. Multiple blocks can be selected.

Optionally, if django-autocomplete-light is installed, the filters will have text-search and autocomplete. DAL support is optional because it brings with it extra static assets, views, and steps to install. wagtailinventory provides system checks on start up that will let users know when that DAL is unavailable, and what to do if they'd like to enable it. These checks are `Info` and `Warning` level, and silenceable.