Remove non-DAL as an option
django-filters has a weird issue where `ChoiceFilter` has a null-option, but its `AllValuesMultipleFilter` can’t resolve that back to the values list when checking for a valid choice.

Rather than propose to fix this upstream (and wait for Wagtail to bring it in), or to fix it in wagtail-inventory (which adds a lot of complexity), I’m dropping no-DAL as an option.

From here django-autocomplete-light is required. This results in less overall code, and less complexity.