Exempt PageBlock from Wagtail's reference index
Wagtail 4.1 introduced a new reference index that keeps track of Django
model instance usage. In versions 4.1 and 4.2, this tracks instances
from any Django model -- in 5.0 this was changed so that only
Wagtail-related models are tracked.

There's no need to track this project's PageBlock model class and, for
sites with a large number of pages and/or blocks, the large number of
PageBlock instances can increase the size and overhead of the reference
index.

This commit exempts the PageBlock model from the index by setting
`wagtail_reference_index_ignore = True`, as per the documentation here:
https://docs.wagtail.org/en/v4.1.6/advanced_topics/reference_index.html