Add ability to limit block inventory permission
Currently any Wagtail user with admin permissions may view the block
inventory report. This commit adds the ability to leverage the
existing wagtailinventory "view_pageblock" permission to grant this
ability to non-admin users.

This permission can be granted in the standard Wagtail group view,
listed as "Can view" for the "Page block" object type.

To test, run an interactive test session with `tox -e interactive`.
First login as the admin user with credentials `admin` / `changeme`.
Confirm that the block inventory exists in the Reports sidebar menu
and that you can view it.

Then, create a new non-admin user at
http://localhost:8000/admin/users/add/. Add this user to the existing
Editors group but do not grant it admin permissions.

Log out of the Wagtail admin and log back in using this new user.
You'll notice this user won't see the block inventory in the Reports
sidebar menu, and if you try to access
http://localhost:8000/admin/block-inventory/ directly, you'll be
redirected with an error message.

Next, log back in as admin, and edit the Editors group at
http://localhost:8000/admin/groups/2/. Give this group the "Can view"
permission for the "Page block" object type. Hit Save.

Finally, log back in as your other user, and confirm that you can now
access the block inventory report.