Update for universal listing report
This change updates the Wagtail Block Inventory report to be a universal listing, per the [changes in Wagtail 6.2](https://docs.wagtail.org/en/latest/releases/6.2.html#changes-to-report-views-with-the-new-universal-listings-ui).

Because the way universal listing filters work doesn't play well with django-autocomplete-lite and Select2, I've removed that dependency.

To make the block selection filters more human-friendly (and less likely to truncate before the actual block name), I've also shortened the block name in the filter to just the block name itself, not the full module path of the block. This does risk some potential duplicate blocks, but I suspect it should be okay.

Finally, because of the switch to universal listings, rather than continue to support old-and-new style reports, I've removed support for Wagtail < 6.2. This will necessitate a major release.